fix: resolve all 76 mypy strict errors across 18 files
- pyproject.toml: add pydantic mypy plugin + jose to ignore_missing_imports
- app/utils/config.py: fix return type (str | None → str)
- app/exchange/client.py: proper dict() returns for ccxt Any values
- app/database.py: add Generator/None return type annotations
- app/bots/schemas.py: replace conint() with Field(ge=, le=) for mypy compat
- app/bots/client.py: typed response.json() return value
- app/bots/router.py: correct type: ignore error codes for OctoBotClient
- app/knowledge/service.py: fix type: ignore codes for optional imports
- app/knowledge/router.py: fix status variable shadowing fastapi.status
- app/auth/service.py: add Any import, typed decode_token return
- app/ai/service.py: handle Anthropic ContentBlock union types,
  handle OpenAI nullable content
- app/automation/background_tasks.py: Task[None] type parameter
- app/automation/scheduled_tasks.py: Task[None] + assert before cancel()
- app/automation/reporting_service.py: Decimal types in build_metrics_summary
- app/automation/router.py: dict[str, Any] return type
- app/aave/client.py: receipt["status"] dict access, type: ignore for
  untyped _get_token_contract
- app/main.py: -> None on all event handlers, dict[str, str] health check,
  lambda returns None for on_health_factor callback

Made-with: Cursor

+++ backend/app/automation/reporting_service.py
@@ -65,9 +65,9 @@ def build_metrics_summary(
     now = now or datetime.now(timezone.utc)
     period_start = now - period
 
-    sums: Dict[str, float] = defaultdict(float)
-    mins: Dict[str, float] = {}
-    maxs: Dict[str, float] = {}
+    sums: Dict[str, Decimal] = defaultdict(Decimal)
+    mins: Dict[str, Decimal] = {}
+    maxs: Dict[str, Decimal] = {}
     counts: Dict[str, int] = defaultdict(int)
 
     for event in events:
@@ -93,9 +93,9 @@ def build_metrics_summary(
     for mid, total in sums.items():
         cnt = counts[mid]
         stats[mid] = {
-            "min": mins[mid],
-            "max": maxs[mid],
-            "avg": total / cnt if cnt else 0.0,
+            "min": float(mins[mid]),
+            "max": float(maxs[mid]),
+            "avg": float(total / cnt) if cnt else 0.0,
             "count": float(cnt),
         }
